TFR model run on ER networks computing centrality
This code solves the TFR model using an Euler scheme on ER graphs and computes graph summary parameters in addition to measures of centrality. This code also saves the time-lapse off the order parameters and performs signal correlation and Fourier analysis. Model parameters and graph summary measures are output to a csv file.