General script running simulation of TFR model
This script takes a csv file containig and adjacency matrix as input. It also takes as input variables for assigning the ratio of switches to oscillators based upon quantiles, as well as whether to implement a phase lag between the switches and the oscillators. It runs a simulation of the TFR model via a euler's method scheme.